Update package.json
In this version, we have added the following:

"dependencies": This section lists the production dependencies required by the project. Here, we have added React and React DOM as dependencies.
More "devDependencies": This section lists the development dependencies required by the project. We have added several dependencies that are commonly used in modern web development, including Babel and ESLint.
"babel": This section provides configuration for Babel, a tool for transpiling modern JavaScript to a backwards-compatible version that can run in older browsers. We have added presets for the env and react plugins.
"eslintConfig": This section provides configuration for ESLint, a tool for enforcing coding standards and detecting potential errors in JavaScript code. We have added configuration that extends the recommended rulesets for ESLint and the react plugin, and turns off the prop-types rule.