Update style.css
In this version, I added some additional styles to the button element to give it some more visual appeal and make it more usable. Specifically, I added a border-radius to make the corners of the button rounded, increased the padding to make the button larger and easier to click/tap, and added a font-size to make the text more legible. I also added a cursor property to change the mouse cursor to a hand icon when hovering over the button, to make it clear that it's clickable.

For the li element, I added some additional styles to make list items more readable and visually appealing. I increased the font-size to make the text easier to read, increased the line-height to add some space between lines, and added a margin-bottom to add some vertical space between list items.