Update details.json
In this version, I've added more detail to the subtitle and description properties to better describe the purpose of the project. I've also updated the files array to include more specific file names and added a new nav property that contains an array of objects representing links to related content. Each object in the nav array contains a label and a link property.